pyproject.toml
[tool.poetry]
name = "texshape"
version = "0.1.0"
description = "TexShape"
authors = ["Kaan Kale"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "2.0.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
pytorch-lightning = "^2.2.3"
scikit-learn = "^1.4.2"
sentence-transformers = "^2.7.0"
datasets = "^2.19.0"
tensorboard = "^2.16.2"
seaborn = "^0.13.2"
emoji = "^2.11.1"
langdetect = "^1.0.9"
matplotlib = "3.8.1"
librosa = "^0.10.2.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
